Add bun refresh script with single-user support
Replace scripts/backfill-profiles.mjs with scripts/refresh.ts, a bun
script mirroring suspend.ts (auto-loads .env, no --env-file). Adds a
<login> mode to refresh one particular user, alongside the existing
default (un-backfilled rows) and --all modes. Documents it in the README.
